Inventory functionality
I added on to the inventory with making changes to inventory and the game.cs file. However I cannot test it because there is issues with fear levels in multiple other rooms due to my editing. I fixed the issue with formatting and the magenta showing up in texts where we didn't want it to be displayed.